requirements.txt
pandas==2.3.1
sqlalchemy==2.0.41
psycopg2-binary==2.9.10
google-cloud-bigquery==3.20.0
selenium==4.34.2
webdriver_manager==4.0.2